fix dry run
