+++ kb_python/utils.py
@@ -1,5 +1,4 @@
 import concurrent.futures
-import functools
 import gzip
 import logging
 import os
@@ -21,7 +20,6 @@ from .config import (
     CHUNK_SIZE,
     get_bustools_binary_path,
     get_kallisto_binary_path,
-    is_dry,
     MAP_DIR,
     PACKAGE_PATH,
     PLATFORM,
@@ -29,6 +27,7 @@ from .config import (
     WHITELIST_DIR,
     UnsupportedOSException,
 )
+from .dry import dryable
 from .dry import utils as dry_utils
 from .stats import STATS
 
@@ -80,33 +79,6 @@ def update_filename(filename, code):
     """
     name, extension = os.path.splitext(filename)
     return f'{name}.{code}{extension}'
-
-
-def dryable(dry_func):
-    """Function decorator to set a function as dryable.
-
-    When this decorator is applied, the provided `dry_func` will be called
-    instead of the actual function when the current run is a dry run.
-
-    :param dry_func: function to call when it is a dry run
-    :type dry_func: function
-
-    :return: wrapped function
-    :rtype: function
-    """
-
-    def wrapper(func):
-
-        @functools.wraps(func)
-        def inner(*args, **kwargs):
-            if not is_dry():
-                return func(*args, **kwargs)
-            else:
-                return dry_func(*args, **kwargs)
-
-        return inner
-
-    return wrapper
 
 
 def open_as_text(path, mode):
