+++ tests/test_utils.py
@@ -13,31 +13,12 @@ from kb_python.config import CHUNK_SIZE, UnsupportedOSException
 from tests.mixins import TestMixin
 
 
-def dry_dummy_function(i):
-    return i + 1
-
-
-@utils.dryable(dry_dummy_function)
-def dummy_function(i):
-    return i
-
-
 class TestUtils(TestMixin, TestCase):
 
     def test_update_filename(self):
         self.assertEqual(
             'output.s.c.bus', utils.update_filename('output.s.bus', 'c')
         )
-
-    def test_dryable_not_dry(self):
-        with mock.patch('kb_python.utils.is_dry') as is_dry:
-            is_dry.return_value = False
-            self.assertEqual(1, dummy_function(1))
-
-    def test_dryable_dry(self):
-        with mock.patch('kb_python.utils.is_dry') as is_dry:
-            is_dry.return_value = True
-            self.assertEqual(2, dummy_function(1))
 
     def test_open_as_text_textfile(self):
         path = os.path.join(
